Never delete the QCoreApplication.
Doing so deletes all posted events without warning.  QueuedFunc is not
set up to handle this, and crashes.  Working around the problem in
QueuedFunc would be extremely messy as it would involve detecting which
events were deleted and reposting them afterwards.  Since there is no
real need to delete the QCoreApplication at present, skip it.