Add an empty last column to the prefs-pluginlist model
This is the column that will be stretched so that the plugin treeview
takes up the full width of the widget that presents it to the user
(which in turn ensures that styles with alternating row colours do not
paint only a fraction of the row). Adding a dedicated "filler" column
means that clicks in that area will not open the settings dialog.