Pass "audacious" to gtk_init as argv[0] so that WM_CLASS gets set correctly
Before this commit the WM_CLASS of gtk2 builds of audacious-3.9
would be set to "<unknown>", "<unknown>".

This commit passes "audacious" to gtk_init as argv[0], which changes
WM_CLASS to "audacious", "Audacious".

This fixes the Window title (in alt-tab) being <unknown> and a non svg icon
being used under GNOME 3.28.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>