Remove inode/directory from the supported MimeTypes.  Reopens: #102.
This should not be necessary and the ONLY reason for doing so is
the broken desktop environments that continue to choose a default
"file manager" based on alphabetical order.  After 6 years users
are (rightfully) still confused why their desktop is broken after
installing Audacious.

This doesn't fix ANYTHING.  The only change will be that the next
application down the list starts getting all the bug reports.
Sorry, EasyTag!