libaudqt: Fix crash when infopopup is being hidden
Hovering over the track I wanted to listen to showed a tool tip (the
InfoPopup widget); if I moved the mouse in any way after the widget was
displayed, it would immediately crash with a segmentation fault.

Valgrind showed that this was due to s_infopopup being deleted from an
event that involved it.  By switching both `delete` to `deleteLater`,
the crash no longer occurs.

Closes: #828.