Fix incorrect tuple for cuesheet entry in some cases.  Closes: #878.
When need_image==true (e.g. scanning an entry for playback), we
would re-read the tuple from the audio file, possibly overwriting
the correct tuple that was read from the cuesheet.

If the entry had already been scanned, this didn't matter, due to
a check in PlaylistData::update_entry_from_scan().  So it was only
noticeable when a single cuesheet entry was opened separately from
the rest of the cuesheet.