audqt: Parse font family names containing digits correctly. Closes: #1483
"PxPlus ToshibaSat 9x16 Bold 8" from the Oldschool PC Font Pack
is such an example. To parse the font family correctly, we need
to check if all characters for the font size are digits.

Both str_to_int() and atoi() do not support this.
Use strtol() instead and check "endptr" accordingly.