Make UI appear on Qt::ApplicationActive for macOS
If the status icon plugin is used with close to system tray the UI main
window is closed. On macOS there is not an easy way to make it reappear.
The system tray activation trigger is disabled on macOS. This is for
good reason due to one button mouse support and that this usually is
intended to open the status/tray menu. Instead we should handle the
application state change activate signal which will fire when the dock
icon is clicked (it abstracts app delegate
applicationShouldHandleReopen).

This was originally fixed with
audacious-media-player/audacious-plugins#174 but it is cleaner to
decouple from any one window/skin implementation via connecting
aud_ui_show to the QApplication directly.