desktop: add directories as supported mime types
Audacious can handle opening a whole directory (in which case in plays
all media inside of it), but won't show up in the "Open With…" options
in file managers (e.g.: nemo) because it doesn't explicitly declare
support for this.

Add inode/directory to supported mime-types so that file managers show
audacious in their "Open With…" menus.